+++ src/model-derivative.ts
@@ -1,4 +1,5 @@
 import { ForgeClient, IAuthOptions, Region } from './common';
+import { isNullOrUndefined } from 'util';
 
 const RootPath = 'modelderivative/v2';
 const ReadTokenScopes = ['data:read'];
@@ -40,7 +41,49 @@ export interface IDerivativeManifest {
     region: string;
     urn: string;
     version: string;
-    //derivatives: any[];
+    derivatives: IDerivative[];
+}
+
+export interface IDerivative {
+    status: string;
+    progress?: string;
+    name?: string;
+    hasThumbnail?: string;
+    outputType?: string;
+    children?: DerivativeChild[];
+}
+
+type DerivativeChild = IDerivativeResourceChild | IDerivativeGeometryChild | IDerivativeViewChild;
+
+export interface IDerivativeChild {
+    guid: string;
+    type: string;
+    role: string;
+    status: string;
+    progress?: string;
+    children?: DerivativeChild[];
+}
+
+export interface IDerivativeResourceChild extends IDerivativeChild {
+    type: 'resource';
+    urn: string;
+    mime: string;
+}
+
+export interface IDerivativeGeometryChild extends IDerivativeChild {
+    type: 'geometry';
+    name?: string;
+    viewableID?: string;
+    phaseNames?: string;
+    hasThumbnail?: string;
+    properties?: any;
+}
+
+export interface IDerivativeViewChild extends IDerivativeChild {
+    type: 'view';
+    name?: string;
+    camera?: number[];
+    viewbox?: number[];
 }
 
 export interface IDerivativeMetadata {
@@ -59,6 +102,54 @@ export enum ThumbnailSize {
     Small = 100,
     Medium = 200,
     Large = 400
+}
+
+/**
+ * Utility class for querying {@see IDerivativeManifest}.
+ */
+export class ManifestHelper {
+    constructor(protected manifest: IDerivativeManifest) {}
+
+    /**
+     * Finds manifest derivatives with matching 'guid', 'type', or 'role' properties.
+     * @param {object} query Dictionary of the requested properties and values.
+     * @returns {DerivativeChild[]} Matching derivatives.
+     */
+    search(query: { guid?: string; type?: string; role?: string; }): DerivativeChild[] {
+        let matches: DerivativeChild[] = [];
+        this.traverse((child: DerivativeChild) => {
+            if ((isNullOrUndefined(query.guid) || child.guid === query.guid)
+                && (isNullOrUndefined(query.type) || child.type === query.type)
+                && (isNullOrUndefined(query.role) || child.role === query.role)) {
+                matches.push(child);
+            }
+            return true;
+        });
+        return matches;
+    }
+
+    /**
+     * Traverses all derivatives, executing the input callback for each one.
+     * @param {(child: DerivativeChild) => boolean} callback Function to be called for each derivative,
+     * returning a bool indicating whether the traversal should recurse deeper in the manifest hierarchy.
+     */
+    traverse(callback: (child: DerivativeChild) => boolean) {
+        function process(node: DerivativeChild, callback: (child: DerivativeChild) => boolean) {
+            const proceed = callback(node);
+            if (proceed && node.children) {
+                for (const child of node.children) {
+                    process(child, callback);
+                }
+            }
+        }
+        for (const derivative of this.manifest.derivatives) {
+            if (derivative.children) {
+                for (const child of derivative.children) {
+                    process(child, callback);
+                }
+            }
+        }
+    }
 }
 
 /**
@@ -165,6 +256,19 @@ export class ModelDerivativeClient extends ForgeClient {
     }
 
     /**
+     * Downloads content of a specific model derivative
+     * ({@link https://forge.autodesk.com/en/docs/model-derivative/v2/reference/http/urn-manifest-derivativeurn-GET/|docs}).
+     * @async
+     * @param {string} modelUrn Model URN.
+     * @param {string} derivativeUrn Derivative URN.
+     * @returns {Promise<ReadableStream>} Derivative content stream.
+     * @throws Error when the request fails, for example, due to insufficient rights, or incorrect scopes.
+     */
+    async getDerivativeStream(modelUrn: string, derivativeUrn: string): Promise<ReadableStream> {
+        return this.getStream(this.region === Region.EMEA ? `regions/eu/designdata/${modelUrn}/manifest/${derivativeUrn}` : `designdata/${modelUrn}/manifest/${derivativeUrn}`, {}, ReadTokenScopes);
+    }
+
+    /**
      * Retrieves metadata of a derivative
      * ({@link https://forge.autodesk.com/en/docs/model-derivative/v2/reference/http/urn-metadata-GET|docs}).
      * @async
@@ -216,5 +320,19 @@ export class ModelDerivativeClient extends ForgeClient {
     async getThumbnail(urn: string, size: ThumbnailSize = ThumbnailSize.Medium): Promise<ArrayBuffer> {
         const endpoint = this.region === Region.EMEA ? `regions/eu/designdata/${urn}/thumbnail` : `designdata/${urn}/thumbnail`;
         return this.getBuffer(endpoint + '?width=' + size, {}, ReadTokenScopes);
+    }
+
+    /**
+     * Retrieves derivative thumbnail stream
+     * ({@link https://forge.autodesk.com/en/docs/model-derivative/v2/reference/http/urn-thumbnail-GET|docs}).
+     * @async
+     * @param {string} urn Document derivative URN.
+     * @param {ThumbnailSize} [size=ThumbnailSize.Medium] Thumbnail size (small: 100x100 px, medium: 200x200 px, or large: 400x400 px).
+     * @returns {Promise<ReadableStream>} Thumbnail data stream.
+     * @throws Error when the request fails, for example, due to insufficient rights, or incorrect scopes.
+     */
+    async getThumbnailStream(urn: string, size: ThumbnailSize = ThumbnailSize.Medium): Promise<ReadableStream> {
+        const endpoint = this.region === Region.EMEA ? `regions/eu/designdata/${urn}/thumbnail` : `designdata/${urn}/thumbnail`;
+        return this.getStream(endpoint + '?width=' + size, {}, ReadTokenScopes);
     }
 }