+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.2.0] - 2019-09-18
+
+### Added
+- Support for streamed upload and download
+- New helper class for searching through Model Derivative manifests
+- Improved typings for Model Derivative manifests and derivatives, so better intellisense!
+- Initial support for parsing SVF files (see src/svf/README.md for details)
+
 ## [5.1.0] - 2019-09-12
 
 ### Added